// src/app/interfaces/user/user.ts
export interface User {
    id?: number;
    firstName: string;
    lastName: string;
    dni: string;
    dniFile: string;
    street: string;
    number: string;
    department: string;
    email: string;
    phone: string;
    mobile: string;
    business: string;
    position: string;
    comercialAddress: string;
    anotherContactFullName: string;
    anotherContactPhone: string;
    anotherContactEmail: string;
    observations: string;
    image: string;
    isActive: boolean;
    customerId: number;
    roleId: number;
    role: any;
    buttons: any;
}